Add separate controller test generator
