+++ package.json
@@ -24,6 +24,7 @@
   ],
   "dependencies": {
     "chalk": "^0.5.0",
+    "lodash": "^3.3.1",
     "path": "^0.11.14",
     "slug": "^0.8.0",
     "util": "^0.10.3",
